requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/base.in
#
aiohttp==3.9.5
    # via discord-py-interactions
aiosignal==1.3.1
    # via aiohttp
amqp==5.2.0
    # via kombu
anyascii==0.3.2
    # via wagtail
anytree==2.12.1
    # via django-modeltree
asgiref==3.8.1
    # via
    #   -r requirements/base.in
    #   channels
    #   daphne
    #   django
    #   django-htmx
attrs==23.2.0
    # via
    #   aiohttp
    #   automat
    #   discord-py-interactions
    #   service-identity
    #   twisted
autobahn==23.6.2
    # via daphne
automat==22.10.0
    # via twisted
beautifulsoup4==4.12.3
    # via
    #   -r requirements/base.in
    #   pynliner
    #   wagtail
billiard==4.2.0
    # via celery
bleach[css]==6.1.0
    # via
    #   bleach
    #   django-markdownfield
    #   django-markdownify
boto3==1.34.130
    # via django-storages
botocore==1.34.130
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.4.0
    # via -r requirements/base.in
certifi==2024.6.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cmarkgfm
    #   cryptography
    #   pygit2
channels[daphne]==4.1.0
    # via -r requirements/base.in
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cmarkgfm==2024.1.14
    # via
    #   -r requirements/base.in
    #   django-tuieditor
constantly==23.10.4
    # via twisted
cryptography==42.0.8
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
cssutils==2.11.1
    # via pynliner
daphne==4.1.2
    # via channels
defusedxml==0.7.1
    # via willow
discord-py-interactions==5.11.0
    # via -r requirements/base.in
discord-typings==0.9.0
    # via discord-py-interactions
dj-stripe==2.8.4
    # via -r requirements/base.in
django==4.2.13
    # via
    #   -r requirements/base.in
    #   channels
    #   dj-stripe
    #   django-admin-csvexport
    #   django-allauth
    #   django-anymail
    #   django-email-log
    #   django-filter
    #   django-htmx
    #   django-markdownfield
    #   django-markdownify
    #   django-modelcluster
    #   django-modeltree
    #   django-multi-email-field
    #   django-permissionedforms
    #   django-recaptcha
    #   django-sesame
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   django-tuieditor
    #   djangorestframework
    #   laces
    #   sentry-sdk
    #   wagtail
    #   wagtail-localize
    #   wagtail-localize-git
django-admin-csvexport==2.2
    # via -r requirements/base.in
django-allauth[mfa]==0.63.3
    # via -r requirements/base.in
django-anymail[mailgun]==11.1
    # via -r requirements/base.in
django-dotenv==1.4.2
    # via -r requirements/base.in
django-email-log==1.4.0
    # via -r requirements/base.in
django-environ==0.11.2
    # via -r requirements/base.in
django-filter==24.2
    # via wagtail
django-htmx==1.18.0
    # via -r requirements/base.in
django-markdownfield==0.11.0
    # via -r requirements/base.in
django-markdownify==0.9.5
    # via -r requirements/base.in
django-modelcluster==6.3
    # via wagtail
django-modeltree==0.5
    # via django-admin-csvexport
django-multi-email-field==0.7.0
    # via -r requirements/base.in
django-ordered-model==3.7.4
    # via -r requirements/base.in
django-permissionedforms==0.1
    # via wagtail
django-recaptcha==4.0.0
    # via -r requirements/base.in
django-sesame==3.2.2
    # via -r requirements/base.in
django-sitetree==1.18.0
    # via -r requirements/base.in
django-storages[s3]==1.14.3
    # via -r requirements/base.in
django-taggit==5.0.1
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
django-tuieditor==0.2.0
    # via -r requirements/base.in
djangorestframework==3.15.2
    # via wagtail
draftjs-exporter==5.0.0
    # via wagtail
emoji==2.12.1
    # via discord-py-interactions
et-xmlfile==1.1.0
    # via openpyxl
filetype==1.2.0
    # via willow
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
geographiclib==2.0
    # via geopy
geopy==2.4.1
    # via -r requirements/base.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wagtail-localize-git
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.7
    # via
    #   hyperlink
    #   requests
    #   twisted
    #   yarl
incremental==22.10.0
    # via twisted
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.7
    # via celery
l18n==2021.3
    # via wagtail
laces==0.1.1
    # via wagtail
mailchimp3==3.0.21
    # via -r requirements/base.in
markdown==3.6
    # via
    #   -r requirements/base.in
    #   django-markdownfield
    #   django-markdownify
more-itertools==10.4.0
    # via cssutils
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==2.0.1
    # via shapely
openpyxl==3.1.4
    # via wagtail
pillow==10.3.0
    # via
    #   -r requirements/base.in
    #   pillow-heif
    #   wagtail
pillow-heif==0.16.0
    # via willow
polib==1.2.0
    # via wagtail-localize
prompt-toolkit==3.0.47
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements/base.in
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.0
    # via service-identity
pycparser==2.22
    # via cffi
pygit2==1.14.1
    # via
    #   -r requirements/base.in
    #   wagtail-localize-git
pynliner==0.8.0
    # via -r requirements/base.in
pyopenssl==24.1.0
    # via twisted
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
pytz==2024.1
    # via
    #   django-modelcluster
    #   l18n
qrcode==7.4.2
    # via django-allauth
redis==5.0.6
    # via celery
requests==2.32.3
    # via
    #   django-anymail
    #   mailchimp3
    #   stripe
    #   wagtail
s3transfer==0.10.1
    # via boto3
sentry-sdk[django]==2.6.0
    # via -r requirements/base.in
service-identity==24.1.0
    # via twisted
shapely==2.0.5
    # via -r requirements/base.in
shortuuid==1.0.13
    # via django-markdownfield
six==1.16.0
    # via
    #   anytree
    #   automat
    #   bleach
    #   l18n
    #   python-dateutil
smmap==5.0.1
    # via gitdb
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.5.0
    # via django
stripe==4.2.0
    # via dj-stripe
telepath==0.3.1
    # via wagtail
tinycss2==1.2.1
    # via bleach
toml==0.10.2
    # via wagtail-localize-git
tomli==2.0.1
    # via discord-py-interactions
twisted[tls]==24.3.0
    # via
    #   daphne
    #   twisted
txaio==23.1.1
    # via autobahn
typing-extensions==4.12.2
    # via
    #   discord-typings
    #   emoji
    #   qrcode
    #   twisted
    #   wagtail-localize
tzdata==2024.1
    # via celery
urllib3==2.2.2
    # via
    #   botocore
    #   django-anymail
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==6.1.2
    # via
    #   -r requirements/base.in
    #   wagtail-localize
    #   wagtail-localize-git
wagtail-localize==1.9
    # via
    #   -r requirements/base.in
    #   wagtail-localize-git
wagtail-localize-git==0.14.1
    # via -r requirements/base.in
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
whitenoise==6.7.0
    # via -r requirements/base.in
willow[heif]==1.8.0
    # via
    #   wagtail
    #   willow
yarl==1.9.4
    # via aiohttp
zope-interface==6.4.post2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
